Added some more tests; fixed bugs mentioned by Guy in review; also eliminate variable if LB==UB is known

+++ src/engine/MaxConstraint.h
@@ -115,7 +115,7 @@ class MaxConstraint : public PiecewiseLinearConstraint
     unsigned _maxElim;
     bool _phaseFixed;
     unsigned _fixedPhase;
-    float _maxLowerBound;
+    double _maxLowerBound;
 
     bool _removePL;
